+++ InventoryControl/Services/Interfaces/IClock.cs
@@ -1,0 +1,6 @@
+namespace InventoryControl.Services.Interfaces;
+
+public interface IClock
+{
+    DateTime UtcNow { get; }
+}