+++ InventoryControl/Validators/Api/CategoryDtoValidator.cs
@@ -1,0 +1,13 @@
+using FluentValidation;
+using InventoryControl.Controllers.Api;
+
+namespace InventoryControl.Validators.Api;
+
+public class CategoryDtoValidator : AbstractValidator<CategoryDto>
+{
+    public CategoryDtoValidator()
+    {
+        RuleFor(x => x.Name).NotEmpty().MaximumLength(100);
+        RuleFor(x => x.Description).MaximumLength(500).When(x => x.Description is not null);
+    }
+}